+++ resources/js/views/settings/users/View.vue
@@ -1,7 +1,5 @@
 <template>
-  <div>
-    <!-- TODO: Implement -->
-  </div>
+
 </template>
 
 <script>